Return tests for external validation rules into `smoke-tests`.
