+++ base/src/test/proto/spine/test/types/known_types_should.proto
@@ -27,7 +27,6 @@ option (type_url_prefix) = "type.spine.io";
 option java_package="io.spine.test.types";
 option java_multiple_files = true;
 
-
 // The types declared below are used to test the `KnownTypes#getTypesFromPackage`.
 //
 // See `KnownTypesShould` for the unit test code.
