+++ tools/smoke-tests/validation-rules/src/test/proto/spine/test/validate/validating_builder_test.proto
